==> analysis_easifish/bydatasets/lt186/14.model_easifish_perturbation.ipynb <==
import numpy as np
import pandas as pd

import matplotlib.pyplot as plt
import seaborn as sns

from scroutines import config_plots #import *

f = '/u/home/f/f7xiesnm/project-zipursky/easifish/results/lt186_240520_abc.csv'
df = pd.read_csv(f)
df['lbl_v2'] = df['lbl_v2'].fillna('NA')
# df = df[df['abc_bin'] =='A']
df

df['abc_bin2'] = pd.cut(df['ptime'], bins=[0,0.56,0.66,0.8, 1], labels=['A', 'B', 'C', 'L4'])
df['abc_depth_bin'] = pd.cut(df['zr'], bins=[100,300,400,450, 600], labels=['A', 'B', 'C', 'L4'])

fig, ax = plt.subplots()
sns.scatterplot(data=df, x='zr', y='ptime', hue='abc_bin', s=5, edgecolor='none', ax=ax)
# ax.invert_yaxis()
plt.show()

fig, ax = plt.subplots()
sns.scatterplot(data=df, x='zr', y='ptime', hue='abc_bin2', s=5, edgecolor='none', ax=ax)
# ax.invert_yaxis()
plt.show()

fig, ax = plt.subplots()
sns.scatterplot(data=df, x='zr', y='ptime', hue='abc_depth_bin', s=5, edgecolor='none', ax=ax)
# ax.invert_yaxis()
plt.show()

fig, ax = plt.subplots()
sns.scatterplot(data=df, x='zr', y='ptime', c='lightgray', s=5, edgecolor='none', ax=ax)
sns.scatterplot(data=df[df['lbl_v2']!='NA'], x='zr', y='ptime', hue='lbl_v2', s=5, edgecolor='none', ax=ax)
# ax.invert_yaxis()
plt.show()

hue = 'abc_bin2'
fig, ax = plt.subplots()
sns.scatterplot(data=df[df[hue].isin(['A', 'B'])], x='xr', y='yr', hue=hue, s=5, edgecolor='none', ax=ax)
# ax.invert_yaxis()
plt.show()

fig, ax = plt.subplots()
sns.scatterplot(data=df[df[hue].isin(['A', 'B'])], x='xr', y='zr', hue=hue, s=5, edgecolor='none', ax=ax)
ax.invert_yaxis()
plt.show()

fig, ax = plt.subplots()
sns.scatterplot(data=df[df[hue].isin(['A', 'B'])], x='yr', y='zr', hue=hue, s=5, edgecolor='none', ax=ax)
ax.invert_yaxis()
plt.show()

fig, ax = plt.subplots()
sns.scatterplot(data=df, x='xr', y='yr', hue='abc_bin2', s=5, edgecolor='none', ax=ax)
# ax.invert_yaxis()
plt.show()

fig, ax = plt.subplots()
sns.scatterplot(data=df, x='xr', y='zr', hue='abc_bin2', s=5, edgecolor='none', ax=ax)
ax.invert_yaxis()
plt.show()

fig, ax = plt.subplots()
sns.scatterplot(data=df, x='yr', y='zr', hue='abc_bin2', s=5, edgecolor='none', ax=ax)
ax.invert_yaxis()
plt.show()

hue = 'abc_depth_bin'
fig, ax = plt.subplots()
sns.scatterplot(data=df, x='xr', y='yr', hue=hue, s=5, edgecolor='none', ax=ax)
ax.invert_yaxis()
plt.show()

fig, ax = plt.subplots()
sns.scatterplot(data=df, x='xr', y='zr', hue=hue, s=5, edgecolor='none', ax=ax)
ax.invert_yaxis()
plt.show()

fig, ax = plt.subplots()
sns.scatterplot(data=df, x='yr', y='zr', hue=hue, s=5, edgecolor='none', ax=ax)
ax.invert_yaxis()
plt.show()

palette_proj = {
    'LM': 'red', 
    'RL': 'blue', 
    'both': 'pink', 
    'NA': 'lightgray', 
}

hue = 'lbl_v2'
fig, ax = plt.subplots()
sns.scatterplot(data=df, x='xr', y='yr', color='lightgray', s=5, edgecolor='none', ax=ax)
sns.scatterplot(data=df[df[hue]!='NA'], x='xr', y='yr', hue=hue, s=5, edgecolor='none', palette=palette_proj, ax=ax)
# ax.invert_yaxis()
plt.show()

fig, ax = plt.subplots()
sns.scatterplot(data=df, x='xr', y='zr', color='lightgray', s=5, edgecolor='none', ax=ax)
sns.scatterplot(data=df[df[hue]!='NA'], x='xr', y='zr', hue=hue, s=5, edgecolor='none', palette=palette_proj, ax=ax)
ax.invert_yaxis()
plt.show()

fig, ax = plt.subplots()
sns.scatterplot(data=df, x='yr', y='zr', color='lightgray', s=5, edgecolor='none', ax=ax)
sns.scatterplot(data=df[df[hue]!='NA'], x='yr', y='zr', hue=hue, s=5, edgecolor='none', palette=palette_proj, ax=ax)
ax.invert_yaxis()
plt.show()

# use ABC to simulate differences

res = df.groupby(['abc_bin2', 'lbl_v2']).size().unstack()[['LM', 'RL', 'both', 'NA']]
res

from scroutines import powerplots

res_frac = res.loc[['A', 'B', 'C'], ['LM', 'RL', 'both']]
res_frac = res_frac.divide(res_frac.sum(axis=1), axis=0)

# colors = ['red', 'blue', 'purple']

fig, ax = plt.subplots(figsize=(2,4))
res_frac.plot.bar(width=0.8, stacked=True, ax=ax, linewidth=0) # , color=colors)
ax.legend(bbox_to_anchor=(1,1))
sns.despine(ax=ax)
ax.grid(False)

powerplots.savefig_autodate(fig, '/u/home/f/f7xiesnm/project-zipursky/easifish/results/fig1.pdf')
plt.show()

resdepth = df.groupby(['abc_depth_bin', 'lbl_v2']).size().unstack()[['LM', 'RL', 'both', 'NA']]
resdepth

n_lm = res.loc['A', 'LM']
n_rl = res.loc['A', 'RL']
n_no = res.loc['A', 'NA']

rs = np.linspace(0,1,100) #[0.1, 0.2, 0.5]
rs_text = [0.1, 0.2, 0.5]

ratio0 = n_lm/n_rl
ratio0

def calc_ratio(r, n_lm, n_rl):
    ratio = n_lm*(1-r)/(n_rl+n_lm*r)
    return ratio

def calc_ratio2(r, p, n_lm, n_rl, n_no):
    n_no2 = n_no - n_no*r*p
    n_lm2 = n_lm - n_lm*r
    n_rl2 = n_rl + n_lm*r + n_no*r*p
    
    ratio = n_lm2/n_rl2
    return ratio

ratios = []
for r in rs:
    ratio = calc_ratio(r, n_lm, n_rl) 
    ratios.append(ratio)
    
ratios_text = []
for r in rs_text:
    ratio = calc_ratio(r, n_lm, n_rl) 
    ratios_text.append(ratio)

fig, ax = plt.subplots()
ax.plot(rs, ratios, '-')
for r in rs_text:
    ax.plot(rs_text, ratios_text, 'o', color='k')
    ratio = calc_ratio(r, n_lm, n_rl)
    ax.text(r, ratio, f'{ratio:.2f}')
    
ax.set_xlabel('conversion rate of the infected cells')
ax.set_ylabel('LM/RL ratio for the infected cells')
plt.show()

ps = np.linspace(0,1,5)
rs = np.linspace(0,1,101) #[0.1, 0.2, 0.5]
texts_idx = np.array([10,20,50])

ratios_p = {}
for p in ps:
    ratios = []
    for r in rs:
        ratio = calc_ratio2(r, p, n_lm, n_rl, n_no) 
        ratios.append(ratio)
    ratios = np.array(ratios)
    ratios_p[p] = ratios
    
fig, ax = plt.subplots()
for p in ps:
    ratios = ratios_p[p]
    ax.plot(rs, ratios, '-', label=f'{p:.2f}')
    # ax.plot(rs[texts_idx], ratios[texts_idx], 'o', color='k')
    # for i in texts_idx:
    #     ax.text(rs[i], ratios[i], f'{ratios[i]:.2f}')
    
ax.legend(title='conversion ratio others/LM (p)', bbox_to_anchor=(1,1))
ax.set_xlabel('conversion rate (r) for LM')
ax.set_ylabel('LM/RL ratio')
sns.despine(ax=ax)
plt.show()
